// src/app/modules/product/product.interface.ts
export interface Variant {
    type: string;
    value: string;
}

export interface Inventory {
    quantity: number;
    inStock: boolean;
}

export interface Product {
    name: string;
    description: string;
    price: number;
    category: string;
    tags: string[];
    variants: Array<Variant>;
    inventory: Inventory;
}